requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
idna==3.10
inotify_simple==1.3.5
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==3.0.2
numpy==2.2.4
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
threadpoolctl==3.6.0
urllib3==2.3.0
Werkzeug==3.1.3
yara-python==4.5.1
